extract `lib.user.lifecycle`
